fix: correct a typo in the description of the --skip_mailer flag
This commit corrects what appears to be a typo in the documentation for the --skip_mailer flag. I wasn't exactly sure what would be best to put here so I based it on the wording of the other flags. If anything this commit is to bring awareness to this typo in a more helpful way than just opening an issue.